Refactor server code
