+++ apps/server/src/types.ts
@@ -1,0 +1,6 @@
+import { Quiz } from '@quiz-lib/core';
+
+export interface Player {
+  id: string;
+  quiz?: Quiz;
+}
